poi_query/requirements.txt
osmnx>=1.3.0
geopandas>=0.12.0
networkx>=3.0
overpy>=0.7
pyyaml>=6.0
scipy>=1.10.0
